Implements saveOpus; Code cleanup;
